04_django/0925/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-appconf==1.0.5
django-imagekit==4.0.2
pilkit==2.0
Pillow==8.3.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
